pyproject.toml
[tool.poetry]
name = "animal-accounting"
version = "0.1.0"
description = "REST API для ведения учета питомцев (собак и кошек)"
authors = ["V-Holodov <w.a.holodov@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.12.4"
flake8-isort = "^4.1.1"
black = "^21.12b0"
Pillow = "^8.4.0"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
